Remove Unit type that was redundant and change gadget package structure.
