+++ actions.go
@@ -1,0 +1,8 @@
+package times
+
+//Actions is an interface that includes a set of Action.
+//Actions would be used in times.Gadget.
+type Actions interface {
+	Add(Action, Time)        //Add adds Action to Actions in designated time.
+	ActionsWhen(Time) Action //ActionsWhen returns Action in designated time.
+}
