+++ gadget/stopwatch/preset.go
@@ -5,9 +5,9 @@ import (
 	"time"
 )
 
-//Standard returns gadget.Stopwatch that implements timeObject.Standard and formatter.Standard adopting minimum unit as second.
-//Hangul returns gadget.Stopwatch that implements timeObject.Standard and formatter.Hangul adopting minimum unit as second.
-//Detail returns gadget.Stopwatch that implements timeObject.Accurate and formatter.Detail adopting minimum unit as millisecond.
+//Standard returns gadget.Stopwatch that implements timeobject.Standard and formatter.Standard adopting minimum unit as second.
+//Hangul returns gadget.Stopwatch that implements timeobject.Standard and formatter.Hangul adopting minimum unit as second.
+//Detail returns gadget.Stopwatch that implements timeobject.Accurate and formatter.Detail adopting minimum unit as millisecond.
 var (
 	Standard = New(time.Second, formatter.Standard())
 	Hangul   = New(time.Second, formatter.Hangul())
