+++ sample/preset/variables.go
@@ -4,19 +4,19 @@ import (
 	"github.com/simp7/times"
 	"github.com/simp7/times/gadget/stopwatch"
 	"github.com/simp7/times/gadget/timer"
-	"github.com/simp7/times/timeObject"
+	"github.com/simp7/times/timeobject"
 )
 
-//GetStandard returns timeObject.Standard object that are set to seconds by parameter sec.
-//Notice: GetStandard is just for sample. We recommend you to use timeObject.Standard function as a substitute.
+//GetStandard returns timeobject.Standard object that are set to seconds by parameter sec.
+//Notice: GetStandard is just for sample. We recommend you to use timeobject.Standard function as a substitute.
 func GetStandard(sec int) times.Time {
-	return timeObject.StandardZero().SetSecond(sec)
+	return timeobject.StandardZero().SetSecond(sec)
 }
 
-//GetAccurate returns timeObject.Accurate object that are set to seconds by parameter sec.
-//Notice: GetAccurate is just for sample. We recommend you to use timeObject.Accurate function as a substitute.
+//GetAccurate returns timeobject.Accurate object that are set to seconds by parameter sec.
+//Notice: GetAccurate is just for sample. We recommend you to use timeobject.Accurate function as a substitute.
 func GetAccurate(sec int) times.Time {
-	return timeObject.AccurateZero().SetSecond(sec)
+	return timeobject.AccurateZero().SetSecond(sec)
 }
 
 //Stopwatches returns an array of gadget.Stopwatch.
